Moved service TYPES identifier to the same file as that of IoC Container
