+++ src/inversify.config.ts
@@ -1,5 +1,4 @@
 import { Container } from "inversify";
-import { TYPES } from "./types/types";
 
 import { IAnswerRepo } from "./interfaces/IAnswerRepo";
 import { ICategoryRepo } from "./interfaces/ICategoryRepo";
@@ -16,6 +15,16 @@ import { QuestionRepo } from "./repositories/QuestionRepo";
 import { QuizRepo } from "./repositories/QuizRepo";
 import { SubmissionRepo } from "./repositories/SubmissionRepo";
 import { UserRepo } from "./repositories/UserRepo";
+
+export const TYPES = {
+    IDifficultyRepo: Symbol.for("IDifficultyRepo"),
+    ICategoryRepo: Symbol.for("ICategoryRepo"),
+    IQuestionRepo: Symbol.for("IQuestionRepo"),
+    IAnswerRepo: Symbol.for("IAnswerRepo"),
+    IUserRepo: Symbol.for("IUserRepo"),
+    IQuizRepo: Symbol.for("IQuizRepo"),
+    ISubmissionRepo: Symbol.for("ISubmissionRepo"),
+};
 
 const container = new Container();
 container.bind<IAnswerRepo>(TYPES.IAnswerRepo).to(AnswerRepo);
