+++ src/middlewares/auth.ts
@@ -4,7 +4,7 @@ import { decodeRefreshToken, decodeToken } from "../jwt/jwt";
 import { TContext } from "../types/TContext";
 import { container } from "../inversify.config";
 import { IUserRepo } from "../interfaces/IUserRepo";
-import { TYPES } from "../types/types";
+import { TYPES } from "../inversify.config";
 
 const TOKEN_BEGIN_INDEX = 7;
