+++ src/repositories/CategoryRepo.ts
@@ -4,7 +4,7 @@ import { Quiz } from "../entity/Quiz";
 import { ICategoryRepo } from "../interfaces/ICategoryRepo";
 import { IQuizRepo } from "../interfaces/IQuizRepo";
 import { container } from "../inversify.config";
-import { TYPES } from "../types/types";
+import { TYPES } from "../inversify.config";
 
 @injectable()
 export class CategoryRepo implements ICategoryRepo {
